refactor(orchestrator): split runtime execution slices
Summary:
- split runtime launcher and runner responsibilities into explicit
  assignment, workspace, process, execution, recovery, and event
  persistence slices while keeping launcher entrypoints stable
- add direct slice-focused tests for ready execution, runtime fact
  reconciliation, interrupt handling, and commentary trace/step writes
- normalize workspace lease timeout cancellation and wait for machine
  websocket metrics in tests so the required validation stays stable

Rationale:
- reduce the amount of launch, execution, and recovery logic growing in
  runtime_launcher.go and runtime_runner.go
- centralize compatibility-oriented commentary and step persistence via
  the event slice instead of scattering it across the main path
- keep CI-aligned orchestrator and httpapi validation deterministic
  under heavier concurrent test load

Tests:
- go test ./internal/orchestrator -run 'TestRuntime(ExecutionSliceStartReadyExecutionsPublishesExecutingLifecycle|EventPersistenceSliceRecordAgentOutputPersistsTraceAndCommentaryStep|RecoverySliceReconcileRuntimeFactsSchedulesContinuationAfterCleanSessionExit|RecoverySliceReconcileInterruptRequestsInterruptsActiveRun)$'
- go test ./internal/orchestrator -run TestWorkspaceInitLeaseManagerHeartbeatRenewsLease -count=3
- go test ./internal/orchestrator -run TestRuntimeLauncherStartRuntimeSessionWaitsForWorkspaceLeaseWithoutConsumingSessionStartTimeout -count=3
- go test ./internal/httpapi -run TestMachineConnectWebsocketPublishesActivityAndMetrics -count=5
- go test ./internal/orchestrator ./internal/httpapi
